AR-backend/requirements/requirements.txt
# AR Live Studio - 依赖配置
# 安装命令: pip install -r requirements.txt

# Web框架
Flask>=2.0
Flask-SocketIO>=5.0
Flask-CORS>=3.0

# 系统监控
psutil>=5.8
prometheus-client>=0.14

# GUI框架
PyQt5>=5.15

# 计算机视觉
opencv-python>=4.5
numpy>=1.20

# HTTP客户端
requests>=2.25
websocket-client>=1.0

# 认证
PyJWT>=2.0

# 邮件支持
python-dotenv>=0.19

# 开发工具
pytest>=6.0
pytest-cov>=2.0

# 性能优化
lru-dict>=1.1

# 可选依赖（用于增强功能）
# 安装命令: pip install -r requirements.txt -r requirements-extra.txt
# face-recognition>=1.3  # 人脸识别（可选）
# dlib>=19.0            # 关键点检测（可选）
# moviepy>=1.0         # 视频处理（可选）
# pyaudio>=0.2         # 音频录制（可选）
